keep old iostream alongside s3stream
